chuyen tu long sang int

+++ BE/src/main/java/com/mycar/nhom13/Repository/CarRepository.java
@@ -1,13 +1,11 @@
 package com.mycar.nhom13.Repository;
+
 import com.mycar.nhom13.Entity.Car;
-
-import java.util.List;
-
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRepository extends JpaRepository<Car,Long>{
+public interface CarRepository extends JpaRepository<Car,Integer>{
     
     Car findByCarId(Long id);
